Fix new set of warnings

Commit migrated from https://github.com/dotnet/coreclr/commit/651789a4a70c46f589109adab17e211d20719e66

+++ src/coreclr/src/pal/tests/palsuite/threading/WaitForSingleObject/WFSOExThreadTest/WFSOExThreadTest.c
@@ -27,7 +27,7 @@ VOID PALAPI APCFunc(ULONG_PTR dwParam);
 DWORD PALAPI WaiterProc(LPVOID lpParameter);
 void WorkerThread(void);
 
-DWORD ThreadWaitDelta;
+int ThreadWaitDelta;
 
 int __cdecl main( int argc, char **argv ) 
 {
@@ -185,7 +185,7 @@ satisfying any threads that were waiting on the object.
         NewTickCount  = 0xFFFFFFFF;
     }
 
-    ThreadWaitDelta = NewTickCount - OldTickCount;
+    ThreadWaitDelta = (int)(NewTickCount - OldTickCount);
 
     ret = CloseHandle(hWaitThread);
     if (!ret)
